feat: Some optimization stuff (#261)
* python benchmark script

* add some methods

* add files option to bin

* utf8

* reorder materialized

* get column names from bind context in explain

* wip more reorder

* recompute table refs for mat scan after reorder

* hash table resize

* things

* split local/global hash table types

* allow operators to produce multiple batches

* initial stats

* column pruning pt 1

* pt 2

* pt 3

* better parquet projectoin

* fix panic

* pass bind context to expr rewrite

* const fold, bump version

* try parsing as i32 first

* bitmap zip

* computed batch module

* selection

* quick fix, hash join not handling selection right

* different storage

* ok

* stuff and things

* physical

* seems reasonable

* ok

* fixup! ok

* german

* german builder

* union metadata

* string is sane

* more

* binary executor

* Uniform

* ternary

* hmm

* ok

* add

* arith

* compile

* concat

* string

* constant value as array

* batch shims

* cast fail state

* cast and try cast technically

* remove concat

* remove slice

* rename sort ops

* select exec

* select

* sure

* fix selection

* fixup! fix selection

* fixup! fix selection

* fixup cross join, delete some coed

* fixup! fixup cross join, delete some coed

* typed/untyped null

* fixup! typed/untyped null

* select the selection

* fixup! select the selection

* rm old compute filter

* rm old compute take code

* rm old scalar

* wip ipc changes

* fixup ipc

* remove lifetime from array data buffer

* fixup row encoding

* rm eval2, select2

* fill state

* fixup sort acc

* fxup formatting

* fixup parquet read/write

* unary update

* fixup! unary update

* from row

* things

* ok

* more

* more

* more

* date part

* rm macros

* decimal special casing

* things

* delete some code

* fixups

* rm array2

* fix selecting selected

* test fixups

* more fixups

* amazing

* fix selecting with nulls

* Preallocate encoded row offsets

* script fixes

* validate decimal precision

* add back case, standard slts passing

* fixup! add back case, standard slts passing

* simplify avg/sum aggregates

* fixup sum unit tests

* add full datatype in error message in extract date part

* replace const eval with const fold rewrite

* fixup ipc test

* lint

* more lint

* more lint

* comment out q15

* fixup! comment out q15

* do a second filter pushdown

+++ crates/rayexec_execution/src/logical/binder/bind_query/bind_select.rs
@@ -194,8 +194,8 @@ mod tests {
 
         let cols = columns_in_scope(&bind_context, bind_context.root_scope_ref());
         let expected = vec![
-            ("?column?".to_string(), DataType::Int64),
-            ("my_alias".to_string(), DataType::Int64),
+            ("?column?".to_string(), DataType::Int32),
+            ("my_alias".to_string(), DataType::Int32),
         ];
 
         assert_eq!(expected, cols);
